refactor: upgrade red dot system to 64-bit FNV-1a hash (v2.0)
- Replace all int pathHash with long (FNV-1a 64-bit) across core, UI, editor, and generated files
- Add composite key routing for dynamic nodes to guarantee zero collision
- Fix ClearNodeRecursive shared-buffer bug; add stackalloc snapshot approach
- Add free-slot recycling in RedDotTrie to prevent unbounded memory growth
- Fix GetActiveTypes allocating overload to delegate to zero-alloc version
- Remove unused Unity.VisualScripting import from RedDotBinder
- Remove redundant RegisterAll call from TestPanel.Awake
- Rewrite README for v2.0: update all API signatures, hash format, and code examples

+++ Assets/Scripts/RedDot/Core/RedDotDataStore.cs
@@ -168,7 +168,7 @@ namespace RedDot
             }
         }
 
-        public RedDotState GetState(int nodeIndex, int pathHash)
+        public RedDotState GetState(int nodeIndex, long pathHash)
         {
             if (nodeIndex < 0 || nodeIndex >= _capacity)
                 return new RedDotState(pathHash, 0, 0, 0);
